Convert the program to a module structure (#29)
