+++ CODEOWNERS
@@ -1,11 +1,12 @@
 # Author(s): Michael Plunkett
-
-./test @chanteriam @michplunkett @ehabich @necabotheking
-./util @chanteriam @michplunkett @ehabich @necabotheking
-
 # People are marked as code owners in the order of: primary_person, secondary_person
 *py @michplunkett @ehabich
 
-./api @chanteriam @necabotheking
-./data_handling @ehabich @chanteriam
-./visualization @necabotheking @michplunkett
+# Author(s): Michael Plunkett
+/reproductive_rights_data_project/test/ @chanteriam @michplunkett @ehabich @necabotheking
+/reproductive_rights_data_project/util/ @chanteriam @michplunkett @ehabich @necabotheking
+
+# Author(s): Michael Plunkett
+/reproductive_rights_data_project/api/ @chanteriam @necabotheking
+/reproductive_rights_data_project/data_handling/ @ehabich @chanteriam
+/reproductive_rights_data_project/visualization/ @necabotheking @michplunkett
