+++ reproductive_rights_data_project/api/abortion_policy_api.py
@@ -5,17 +5,18 @@ APIs and saving that data to the `../data` folder.
 
 import requests
 import os
-from util.constants import (
+from reproductive_rights_data_project.util.constants import (
     FILE_NAME_ABORTION_POLICY_API_GESTATION,
     FILE_NAME_ABORTION_POLICY_API_INSURANCE,
     FILE_NAME_ABORTION_POLICY_API_MINORS,
     FILE_NAME_ABORTION_POLICY_API_WAITING_PERIOD,
+    FILE_NAME_STATE_NAMES,
     FILTERED_CHARACTERS_REGEX,
     TYPE_DEFAULTS,
     STANDARD_ENCODING,
 )
-from util.constants import REQUEST_TIMEOUT
-from util.functions import to_json
+from reproductive_rights_data_project.util.constants import REQUEST_TIMEOUT
+from reproductive_rights_data_project.util.functions import to_json
 import re
 
 # Author(s): Chanteria Milner, Michael Plunkett
@@ -109,7 +110,7 @@ def clean(state_policies):
     # Get list of states for data cleaning
     states = []
 
-    with open("data/states.txt", "r", encoding=STANDARD_ENCODING) as f:
+    with open(FILE_NAME_STATE_NAMES, "r", encoding=STANDARD_ENCODING) as f:
         for state in f:
             states.append(state.strip())
