+++ reproductive_rights_data_project/data_handling/ansirh/process.py
@@ -3,9 +3,16 @@ Process ANSIRH data into JSON files binned by state.
 """
 
 import pandas as pd
-from data_handling.ansirh.clean import clean
-from util.constants import FILE_NAME_ANSIRH_CLEAN_DATA
-from util.functions import translate_code_to_state, to_json
+from reproductive_rights_data_project.data_handling.ansirh.clean import (
+    clean,
+)
+from reproductive_rights_data_project.util.constants import (
+    FILE_NAME_ANSIRH_CLEAN_DATA,
+)
+from reproductive_rights_data_project.util.functions import (
+    translate_code_to_state,
+    to_json,
+)
 
 
 def clean_and_save():
@@ -19,7 +26,9 @@ def clean_and_save():
         Writes JSON file with cleaned and formatted ANSIRH data.
     """
     # TODO: Will refactor with csv in the near future
-    ansirh_data = pd.read_csv("./data/AFD_2021_for_ArcGIS_Upload.csv")
+    ansirh_data = pd.read_csv(
+        "reproductive_rights_data_project/data/AFD_2021_for_ArcGIS_Upload.csv"
+    )
 
     # Drop empty column
     ansirh_data = ansirh_data.drop(["Unnamed: 2"], axis=1)
