+++ reproductive_rights_data_project/data_handling/service.py
@@ -2,7 +2,9 @@
 This file works as the central point for running the data_handling package.
 """
 
-from data_handling.ansirh.process import clean_and_save
+from reproductive_rights_data_project.data_handling.ansirh.process import (
+    clean_and_save,
+)
 
 
 def main():
