+++ reproductive_rights_data_project/visualization/app.py
@@ -8,10 +8,9 @@ from dash import Dash, html, dcc, Input, Output
 import dash_bootstrap_components as dbc
 
 # import the abstract classes for the visualizations
-from visualization.functions import read_state
-from visualization.maps.usa_country import USAMap as USAMap
-from visualization.maps.usa_state import USAState as USAState
-from visualization.charts.state_summary import StateSummary as StateSummary
+from reproductive_rights_data_project.visualization.functions import (
+    read_state,
+)
 
 DASH_INSTANCE = Dash(__name__, external_stylesheets=[dbc.themes.LUX])
