+++ reproductive_rights_data_project/visualization/functions.py
@@ -3,11 +3,11 @@ This file contains helper functions for generating data visualizations.
 """
 
 import json
-from util.constants import FILE_NAME_ANSIRH_CLEAN_DATA, STANDARD_ENCODING
-
-# Maybe not good for testability
-with open(FILE_NAME_ANSIRH_CLEAN_DATA, encoding=STANDARD_ENCODING) as f:
-    NATIONAL_CLINICS = json.load(f)
+from reproductive_rights_data_project.util.constants import (
+    FILE_NAME_ANSIRH_CLEAN_DATA,
+    FILE_NAME_STATE_NAMES,
+    STANDARD_ENCODING,
+)
 
 
 def get_state_clinic_counts():
@@ -19,6 +19,10 @@ def get_state_clinic_counts():
     Returns:
         (dict): state: clinic count
     """
+
+    # Maybe not good for testability
+    with open(FILE_NAME_ANSIRH_CLEAN_DATA, encoding=STANDARD_ENCODING) as f:
+        NATIONAL_CLINICS = json.load(f)
 
     count_state_clinics = {}
 
@@ -40,6 +44,10 @@ def get_city_clinic_counts():
     Returns:
         (dict): zipcode: clinic count
     """
+
+    # Maybe not good for testability
+    with open(FILE_NAME_ANSIRH_CLEAN_DATA, encoding=STANDARD_ENCODING) as f:
+        NATIONAL_CLINICS = json.load(f)
 
     count_city_clinics = {}
     for state, zipcodes in NATIONAL_CLINICS.items():
@@ -67,6 +75,10 @@ def get_zipcode_clinic_counts():
     Returns:
         (dict): zipcode: clinic count
     """
+
+    # Maybe not good for testability
+    with open(FILE_NAME_ANSIRH_CLEAN_DATA, encoding=STANDARD_ENCODING) as f:
+        NATIONAL_CLINICS = json.load(f)
 
     count_zipcode_clinics = {}
     for _, zipcodes in NATIONAL_CLINICS.items():
@@ -109,6 +121,6 @@ def read_state():
         (list) list of states in alphabetical order
     """
 
-    with open("./data/states.txt", encoding=STANDARD_ENCODING) as state:
+    with open(FILE_NAME_STATE_NAMES, encoding=STANDARD_ENCODING) as state:
         state_lst = [line.strip() for line in state.readlines()]
     return state_lst
