--- util/constants.py
@@ -1,29 +1,0 @@
-"""
-Constants to be used in various parts of the application.
-"""
-
-# Misc. Variables
-# Author(s): Chanteria Milner, Michael Plunkett
-FILTERED_CHARACTERS_REGEX = r'[!.,\'"?:<>]'
-TYPE_DEFAULTS = {"str": None, "bool": False, "int": 0, "float": 0.0}
-STANDARD_ENCODING = "utf-8"
-
-# File Names
-# Author(s): Michael Plunkett
-FILE_NAME_ABORTION_POLICY_API_GESTATION = (
-    "data/abortion_policy_api_gestational.json"
-)
-FILE_NAME_ABORTION_POLICY_API_INSURANCE = (
-    "data/abortion_policy_api_insurance.json"
-)
-FILE_NAME_ABORTION_POLICY_API_MINORS = "data/abortion_policy_api_minors.json"
-FILE_NAME_ABORTION_POLICY_API_WAITING_PERIOD = (
-    "data/abortion_policy_api_waiting.json"
-)
-FILE_NAME_ANSIRH_CLEAN_DATA = "data/clean_ansirh.json"
-
-FILE_NAME_STATE_ABBREVIATIONS = "data/state_abbreviations.csv"
-
-# API Variables
-# Author(s): Michael Plunkett
-REQUEST_TIMEOUT = 10
